requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
faststream[rabbit,kafka]>=0.5.0
grpcio>=1.68.0
grpcio-tools>=1.68.0
loguru>=0.7.0
fire>=0.7.0
orjson>=3.10.0
pydantic-settings>=2.0.0
supervisor>=4.2.0
pyjwt>=2.8.0
